+++ robot_descriptions/booster_t1_description.py
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2025 Inria
 
-"""PiPER MJCF description."""
+"""Booster T1 description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -12,10 +12,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "mujoco_menagerie",
+    "booster_gym",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "agilex_piper")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "resources", "T1")
 
-MJCF_PATH: str = _path.join(PACKAGE_PATH, "piper.xml")
+URDF_PATH: str = _path.join(PACKAGE_PATH, "T1_serial.urdf")