+++ robot_descriptions/n1_description.py
@@ -1,10 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-#
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2025 Inria
 
-"""PiPER MJCF description."""
+"""Fourier N1 description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -12,10 +9,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "mujoco_menagerie",
+    "Wiki-GRx-Models-FourierN1",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "agilex_piper")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "N1")
 
-MJCF_PATH: str = _path.join(PACKAGE_PATH, "piper.xml")
+URDF_PATH: str = _path.join(PACKAGE_PATH, "urdf", "N1_raw.urdf")