+++ robot_descriptions/wl_p311d_description.py
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2025 Inria
 
-"""PiPER MJCF description."""
+"""WL P311D description."""
 
 from os import getenv as _getenv
 from os import path as _path
@@ -12,10 +12,10 @@ from os import path as _path
 from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache(
-    "mujoco_menagerie",
+    "limxdynamics_robot-description",
     commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
 )
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "agilex_piper")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "wheellegged", "WL_P311D")
 
-MJCF_PATH: str = _path.join(PACKAGE_PATH, "piper.xml")
+URDF_PATH: str = _path.join(PACKAGE_PATH, "urdf", "robot.urdf")